Support dlna profile ids for AlbumArtUris

+++ src/upnpavtransportservice.cpp
@@ -87,7 +87,7 @@ ActionResponse Service::onAction(const std::string& action, const xml::Document&
     {
         ActionResponse response(action, ServiceType::AVTransport);
         auto req = doc.getFirstChild();
-        uint32_t id = std::stoul(req.getChildNodeValue("InstanceID"));
+        uint32_t id = static_cast<uint32_t>(std::stoul(req.getChildNodeValue("InstanceID")));
         
         switch (actionFromString(action))
         {
